requirements.txt
aniso8601==9.0.1
click==8.1.3
Flask==2.2.3
Flask-RESTful==0.3.9
flask-restful-swagger-2==0.35
Flask-SQLAlchemy==3.0.3
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.1
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==1.4.32
Werkzeug==2.2.3
zipp==3.15.0
